+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [0.0.2 UNRELEASED]
 
+### Added
+
+- Using Vault to generate username and password for postgres with kv #13
+
 ## [0.0.1]
 
 ### Added